fix(sdk): tolerate unsupported Windows directory fsync in broker lifecycle and coordinator MCP state
On Windows, fsync on a directory handle fails with EPERM (libuv limitation).
Four durable-state paths fsynced a parent directory after an atomic temp-file
rename without the win32 guard that the workflow-gate store already has:

- sdk/broker/lifecycle.ts syncDirectory (effect-marker write): the failure
  fired AFTER a successful session spawn, so the broker terminated the
  freshly spawned healthy session and reported terminal_uncertain, breaking
  session.create via the SDK broker, Telegram /session_create, and
  coordinator MCP start_session on Windows.
- sdk/broker/lifecycle-ledger.ts #syncDirectory (compaction)
- coordinator-mcp/question-state.ts fsyncDirectory (writeAtomic)
- coordinator-mcp/server.ts writeCoordinatorIdempotencyFile

Extract the classifier shipped in the workflow-gate store fix into
src/utils/directory-sync.ts (re-exported from workflow-gate-broker.ts for
compatibility) and apply it at the four sites. Only classified codes
(EINVAL/ENOTSUP/EOPNOTSUPP/EPERM on win32) are tolerated during directory
sync; open() failures, file fsync failures, and all other errors remain
fail-closed. No behavior change on Linux/macOS.
