+++ packages/coding-agent/src/modes/shared/agent-wire/workflow-gate-broker.ts
@@ -14,6 +14,7 @@
 import { closeSync, fsyncSync, mkdirSync, openSync, readFileSync, renameSync, writeFileSync } from "node:fs";
 import * as path from "node:path";
 import type { AskSelectedAckOutcome } from "../../../tools";
+import { isUnsupportedWindowsDirectorySyncError } from "../../../utils/directory-sync";
 import { classifyAskGateDisposition } from "./deep-interview-gate";
 import { answerHashOf, canonicalJson, compileGateSchema, schemaHash, validateGateAnswer } from "./workflow-gate-schema";
 import type {
@@ -970,15 +971,7 @@ function assertPersistedGate(
 		invalidFileState(filePath);
 }
 
-export function isUnsupportedWindowsDirectorySyncError(
-	error: unknown,
-	platform: NodeJS.Platform = process.platform,
-): boolean {
-	if (platform !== "win32") return false;
-	const code = (error as NodeJS.ErrnoException | undefined)?.code;
-	// Bun reports EPERM when fsyncSync is applied to a Windows directory handle.
-	return code === "EINVAL" || code === "ENOTSUP" || code === "EOPNOTSUPP" || code === "EPERM";
-}
+export { isUnsupportedWindowsDirectorySyncError };
 
 /** A write may have reached `rename` but not the directory fsync durability barrier. */
 export class GateStoreWriteError extends Error {
