+++ packages/coding-agent/src/sdk/broker/lifecycle-ledger.ts
@@ -2,6 +2,7 @@ import { createHash } from "node:crypto";
 import * as fsSync from "node:fs";
 import * as fs from "node:fs/promises";
 import path from "node:path";
+import { isUnsupportedWindowsDirectorySyncError } from "../../utils/directory-sync";
 import type { SdkStartupFailure, SdkStartupRollbackResult } from "../startup-capability";
 import { parseLifecycleJson } from "./lifecycle-codec";
 import { assertSupportedStateVersion, SDK_STATE_VERSION } from "./state-version";
@@ -426,6 +427,9 @@ export class LifecycleLedger {
 		const directory = await fs.open(path.dirname(this.#file), fsSync.constants.O_RDONLY);
 		try {
 			await directory.sync();
+		} catch (error) {
+			// Windows cannot fsync a directory handle; rename durability is best-effort there.
+			if (!isUnsupportedWindowsDirectorySyncError(error)) throw error;
 		} finally {
 			await directory.close();
 		}
